move REST API endpoints to JITM package
