+++ projects/packages/jitm/src/class-jitm.php
@@ -66,6 +66,8 @@ class JITM {
 			return false;
 		}
 
+		add_action( 'rest_api_init', array( __NAMESPACE__ . '\\Rest_Api_Endpoints', 'register_endpoints' ) );
+
 		add_action( 'current_screen', array( $this, 'prepare_jitms' ) );
 		return true;
 	}
